requirements.txt
Django==4.0.4
django_environ==0.8.1
djangorestframework==3.13.1
djangorestframework_simplejwt==5.1.0
environ==1.0
kafka==1.3.5
